[CREATE] - Generate Slack Exception Service
