--- src/main/kotlin/sinhee/kang/tutorial/infra/api/slack/SlackSenderManager.kt
@@ -1,67 +1,0 @@
-package sinhee.kang.tutorial.infra.api.slack
-
-import com.fasterxml.jackson.core.JsonProcessingException
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.PropertyNamingStrategy
-import okhttp3.*
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.scheduling.annotation.Async
-import org.springframework.stereotype.Component
-import org.springframework.web.context.request.RequestContextHolder
-import org.springframework.web.context.request.ServletRequestAttributes
-import sinhee.kang.tutorial.global.error.ErrorHandler
-import sinhee.kang.tutorial.infra.api.slack.dto.Attachment
-import sinhee.kang.tutorial.infra.api.slack.dto.Field
-import sinhee.kang.tutorial.infra.api.slack.dto.SlackMessageRequest
-import javax.servlet.http.HttpServletRequest
-
-@Component
-class SlackSenderManager(
-    @Value("\${web.hook.url}")
-    private val webHookUrl: String
-) {
-    private val errorHandler = ErrorHandler()
-    private val client = OkHttpClient()
-    private val mapper = ObjectMapper()
-
-    @Async
-    fun send(request: HttpServletRequest, exception: Exception) {
-        val attachment: SlackMessageRequest = this.toAttachments(request, exception)
-        val sendRequest: Request = Request.Builder()
-                .url(webHookUrl)
-                .addHeader("Content-Type", "application/json")
-                .post(writeValueAsRequestBody(attachment))
-                .build()
-        client.newCall(sendRequest).execute().body()!!.string()
-    }
-
-    fun toAttachments(request: HttpServletRequest, exception: Exception): SlackMessageRequest {
-        val attachmentRequest: MutableList<Attachment> = ArrayList()
-        val fieldList: MutableList<Field> = ArrayList()
-        fieldList.add(Field(title = "Request URI", value = request.requestURI))
-        fieldList.add(Field(title = "Request Header", value = errorHandler.getHeader(request)))
-        fieldList.add(Field(title = "Request Body", value = errorHandler.getBody(request)))
-
-        attachmentRequest.add(Attachment(
-                color = "#FF4444",
-                pretext = "[Server RuntimeError Report]",
-                tile = exception.localizedMessage,
-                text = errorHandler.getStackTrace(exception),
-                footer = "Bug Reporter",
-                ts = System.currentTimeMillis() / 1000,
-                fields = fieldList
-        ))
-
-        return SlackMessageRequest(attachments = attachmentRequest)
-    }
-
-    fun writeValueAsRequestBody(obj: SlackMessageRequest): RequestBody {
-        return try {
-            mapper.propertyNamingStrategy = PropertyNamingStrategy.SNAKE_CASE
-            RequestBody.create(MultipartBody.FORM, mapper.writeValueAsString(obj))
-        }
-        catch (e: JsonProcessingException) {
-            throw IllegalArgumentException(e.message)
-        }
-    }
-}
