+++ src/main/kotlin/sinhee/kang/tutorial/infra/api/slack/service/SlackExceptionService.kt
@@ -1,0 +1,7 @@
+package sinhee.kang.tutorial.infra.api.slack.service
+
+import javax.servlet.http.HttpServletRequest
+
+interface SlackExceptionService {
+    fun sendMessage(request: HttpServletRequest, exception: Exception)
+}
